Handle command batch in JSON format (#8)
* Handle command batch in JSON format

This commit adds an `Exec` trait which is autoimplemented for
majority of intended use-cases. `Exec`, much like `HandleCmds`
executes a sequence of commands returning an async `Stream` in
the process, however, with this difference that `Exec` accepts
a JSON `String` as an argument. Typical use case would involve
a batch like the following:

```
[
    { "start": {}},
    { "transfer" : { "from": "", "to": "" }}
]
```

In case that the container receives a command it doesn't understand,
an error is returned which _does not_ terminate processing of the
rest of the supplied commands.

This commit also updates the `dummy` exe unit to use the `Exec`
trait.

* Fix formatting