Add simple state machine (unoptimised)
This commit represents the internal workings of DummyExeUnit in the
form of a simplistic state machine. _Note_ that the state machine
is just a PoC and definitely unoptimised.

This commit also swaps `std::sync::Mutex` for `futures::lock::Mutex`
which is more suitable for working with futures and reactors.

This commit also adds missing commands to `DummyExeUnit` which now
total 5: deploy, start, run, transfer, and stop.

Finally, this commit introduces a few basic unit tests for the
`DummyExeUnit`.