(Partially) handle CtrlC event in DummyExeUnit
This commit adds partial support for CtrlC events (cross-platform)
which will exit the main loop when such an event occurs. The support
is partial since `std::io::{stdin, stdout}` is still blocking, hence
sometimes it is required to hit a combination of Ctrl-C followed by
an Enter to quit the interactive mode.