Payment API database model change (contd.)
Single node can host multiple identities. It is possible that one of
them being provider and the other requestor, a payment document is
stored in the same db twice. To avoid ID clashes `owner_id` was added
to the database models and included in primary keys.

In most tables `owner_id` is accompanied by `role` (either `R` or `P`)
indicating whether the owner owns an entity as requestor or as
provider. `role` is not included in primary keys because there are no
valid scenarios in which a single identity owns a document both as
requestor and provider.

`invoice_status` renamed to `document_status` because it is used by
both invoices and debit notes.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>