Payment API: Accept debit note (dummy implementation)
Debit notes can be accepted now. It results in setting debit note
status to `ACCEPTED` on both provider and requestor nodes, as well
as emitting an event on the provider node.

There is **no checking** if the allocation used to accept the debit
note has sufficient funds and the payment is **not scheduled**.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>