Payment driver API
Rationale
----

New API introduced for the communication between payment service and
payment driver(s). It is meant to enable the implementation of the
following requirements:
1. Supporting multiple accounts and payment drivers on a single yagna
   daemon.
2. Supporting multiple payment platforms (e.g. GNT, NGNT, NGNT-zkSync).
3. Allowing third-party developers to implement their own payment
   drivers with ease and possibly much freedom.

Detailed list of changes
----

`BUS_ID` is replaced with `BUS_ID_PREFIX` because each particular driver
will have its own ID.

Currency enum is abandoned. It is not desired to have a fixed set of
available currencies. It is assumed that all amounts and balances are
given in currency appropriate for the platform.

All gas-related parameters are abandoned. They were too driver-specific
to be included in the general API. It is recommended that payment
drivers which use gas allow to configure limits (e.g. by env variables).

`SchedulePayment` does not include `invoice_id` parameter anymore.
Instead, it returns an ID of the payment order generated by the driver.

`GetPaymentStatus` is removed because the new payment driver API relies
on push notifications instead. That is realized by the `NotifyPayment`
endpoint added to the payment service. It shall be called by a driver
after every successful transaction.

`Init` message is removed from payment API (GSB). The init command from
CLI will accept `driver` parameter and route the call directly to the
appropriate driver.

`RegisterAccount` / `UnregisterAccount` messages are added because the
payment service needs to maintain a list of available accounts. Payment
drivers should call `RegisterAccount` after every successful
initialization of an account to make it 'visible' to the payment
service. `UnregisterAccount` should be called when the driver is being
shut down. The list of available accounts shall not be persisted by the
payment service. If an account doesn't need to be re-initialized after
restarting yagna daemon, then payment driver should store this
information and call `RegisterAccount` immediately after startup.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>