Payment: initialize accounts from file
On startup Yagna daemon will automatically initialize all payment
accounts listed in `accounts.json` file (path to this file can be
changed using `ACCOUNT_LIST` variable). If the file does not exist, it
will be automatically created with the default identity listed as
receive-only GNT account. Format of the file is specified by an example
`accounts-example.json`.

A new command was added to the CLI in order to verify if the
initialization is working correctly: `yagna payment accounts`. This
command lists all initialized payment accounts with information about
platform, driver, and mode.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>