Payment driver API: added platform
Added `platform` field to `PaymentDetails` and `SchedulePayment`.
Supporting multiple networks invalidates the assumption that every
driver supports a single payment platform. Therefore, this additional
information is needed to be able to make appropriate payment.

Implemented validation of payment platforms in PaymentProcessor to
secure against cross-network replay attacks (i.e. making payment on
testnet and claiming it to be a mainnet payment).

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>