DNS resolve in faucet requests
Due to a bug in `awc` crate default DNS resolution is not working
correctly on Windows, leading to request timeouts. A function for
'manual' resolution has been used for this in ERC-20 but zkSync
driver lacked this.

DNS-resolving function has been moved to `ya_utils_networking` and
is used now in both drivers.