Payment service: fixed race condition in PaymentProcessor
If `NotifyPayment` got called during processing of `SchedulePayment`
(which happened when running `debit_note_flow` example with dummy
driver) it would fail because the payment order had not been saved in
the database yet.

To avoid this race condition (and possibly others) PaymentProcessor
has been wrapped in a mutex. This allowed to remove the mutex from
DriverRegistry. Also, sending `SendPayment` message needs to be done
via `Arbiter::spawn` now to avoid deadlock in case it is sent to the
same node.

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>