ZkSync: Check balance before operations
Check account balance before making a zkSync account unlock or
withdrawal. This allows to display better error messages to the user and
avoid panic in the `withdraw` function (`balance` and `withdraw_fee` are
of `BigUint` type so `balance - withdraw_fee` would cause panic in case
the fee is bigger than the account balance).

Signed-off-by: Adam Wierzbicki <awierzbicki@golem.network>