Fix CI failing since rustc 1.65.0
Two issues.

1. Backtrace
BT format changed in Rust 1.65, causing
bactrace_util::generate_backtraced_name to return an autogenerated name
created via uuidv4. However, changes applied since then have actually
made the happy path a hard requirement by expecting the resultant string
to be of the form r"(.*)::(.*)::.*". This was the case when the test
names were pulled from BT, but didn't work for autogenerated uuidv4 ones.

This patch removed the uuid name generation from backtrace_util, and
implemented it in mock_node in a different manner.

2. split-debuginfo=unpacked fails on Windows.
Cargo should not forward this option if not supported, but this does not
work in this case. Likely a regression in cargo-rustc interaction. For
now removed this flag with an appropriate remark to get the CI going.