Allow payment init --sender with 0 funds
Previously a sender couldn't be initialized with no funds, but this
mechanism doesn't make much sense anymore as we default-initialize
accounts as senders when needed. This meant that payment fund would fail
due to lacking funds, which is undesirable.

Parsing wallet address and checking if GLM/ETH funds can be inspected is
retained to catch other failures.