gsb-api: fix test_json
This test used a serde_json::Value -> flexbuffers -> Rust struct serde
trip which is not guaranteed to work and should not be relied upon. The
proper route is Value -> flexbuffers -> Value -> Rust.

This only broke recently because without the arbitrary_precision feature
of serde_json the shorter path works. The feature has been enabled due
to the addition of erc20_payment_lib to the dependency graph, which
transitively depends on serde_json with arbitrary prec. Note that cargo
unifies duplicate dependencies under the assumption that features are
purely additive, hence the spooky action at a distance.